refactor: TaskSettingView로 뷰 분리 작업
